float

+++ src/utils.py
@@ -2,7 +2,6 @@ import pyaudio
 import wave
 import numpy as np
 from scipy.fft import fft
-from torch import float32
 #import soundfile as sf
 #import librosa as lb
 #import scipy as sc
@@ -33,7 +32,7 @@ class record_audio():
         for ii in range(0,int((self.samp_rate/self.chunk)*self.record_secs)):
             data = self.stream.read(self.chunk)
             
-            data_float = np.fromstring(data, dtype = np.float32)
+            data_float = np.fromstring(data, dtype = np.float)
             if ii == 1:
                 print(data_float)
             #frame2.append(np.fft.fft(np.fft.fft(data_float, n = int(self.chunk/2 + 1)), n = int(self.chunk)))
